requirements/requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
nltk==3.4.5
numpy==1.21.6
pandas==1.3.5
Pillow==9.4.0
pydicom==2.3.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.00
